--- Add Code Here/tree/.vscode/settings.json
@@ -1,3 +1,0 @@
-{
-    "C_Cpp.errorSquiggles": "Disabled"
-}
