--- Add Code Here/tree/tempCodeRunnerFile.cpp
@@ -1,1 +1,0 @@
- cout<<temp->data<<" ";
