--- Contact Html/index.html
@@ -1,24 +1,0 @@
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Contact Form</title>
-</head>
-<body>
-    <h1> Contact Form </h1>
-    <form action="submit">
-        <h2> First Name : 
-            <input id="Fname" class="input" name="Fname" type="text" value="" size="30" /><br />
-        </h2>
-        <h2> Last Name : 
-            <input id="Lname" class="input" name="Lname" type="text" value="" size="30" /><br />
-        </h2>
-        <h2> Mobile No : 
-            <input id="mobile" class="input" name="mobile" type="number" value="" size="30" /><br />
-        </h2>
-    </form>
-    <h2><button type="submit">Submit</button></h2>
-    <script src="./script.js"></script>
-</body>
-</html>
